Refactor app.py to use lifespan context manager for database initialization; add schema.py for Pydantic models
